+++ packages/extension-contracts/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./activation-events.js";
+export * from "./dependencies.js";
 export * from "./ids.js";
 export * from "./manifest-json-schema.js";
 export * from "./manifest.js";